requirements.txt
tensorflow>=2.0.0
tensorflow-addons==0.6.0
numpy>=1.16
pyunsplash==1.0.0b8
pillow==6.1.0
ortools==7.3.7083
numba==0.46.0
scikit-image==0.16.2
